requirements.txt
Discord.py>=2.3.2,<3.0
aiosqlite>=0.20.0,<1.0
python-dotenv>=1.0.0,<2.0
aiohttp>=3.9.0,<4.0
PyNaCl>=1.5.0,<2.0
yt-dlp>=2024.0.0
Pillow>=10.0.0,<12.0
PyPDF2>=3.0.0,<4.0
